+++ opendrive/src/lane/road_mark/sway.rs
@@ -7,18 +7,18 @@ use serde::{Deserialize, Serialize};
 pub struct Sway {
     /// Polynom parameter a, sway value at @s (ds=0)
     #[serde(rename = "@a")]
-    a: f32,
+    a: f64,
     /// Polynom parameter b
     #[serde(rename = "@b")]
-    b: f32,
+    b: f64,
     /// Polynom parameter c
     #[serde(rename = "@c")]
-    c: f32,
+    c: f64,
     /// Polynom parameter d
     #[serde(rename = "@d")]
-    d: f32,
+    d: f64,
     /// s-coordinate of start position of the `<sway>` element, relative to the @sOffset given in
     /// the `<roadMark>` element
     #[serde(rename = "@d_s")]
-    d_s: f32,
+    d_s: f64,
 }
