+++ opendrive/src/lane/type_link.rs
@@ -12,7 +12,7 @@ pub struct TypeLine {
     pub color: Option<Color>,
     /// Length of the visible part
     #[serde(rename = "@length")]
-    pub length: f32,
+    pub length: f64,
     /// Rule that must be observed when passing the line from inside, for example, from the lane
     /// with the lower absolute ID to the lane with the higher absolute ID
     #[serde(rename = "@rule")]
@@ -20,15 +20,15 @@ pub struct TypeLine {
     /// Initial longitudinal offset of the line definition from the start of the road mark
     /// definition
     #[serde(rename = "@sOffset")]
-    pub s_offset: f32,
+    pub s_offset: f64,
     /// Length of the gap between the visible parts
     #[serde(rename = "@space")]
-    pub space: f32,
+    pub space: f64,
     /// Lateral offset from the lane border.
     /// If `<sway>` element is present, the lateral offset follows the sway.
     #[serde(rename = "@tOffset")]
-    pub t_offset: f32,
+    pub t_offset: f64,
     /// Line width
     #[serde(rename = "@width")]
-    pub width: Option<f32>,
+    pub width: Option<f64>,
 }
